requirements.txt
python-dotenv>=1.0.1
requests>=2.32.0
gspread>=6.2.1
google-auth>=2.40.3
openai>=1.0.0




